+++ src/main/java/com/codenal/calendar/domain/CalendarDto.java
@@ -22,7 +22,7 @@ public class CalendarDto {
 	private String calendar_schedule_title;
 	private String calendar_schedule_location;
 	private String calendar_schedule_content;
-	private int calendar_schedule_writer;
+	private Long calendar_schedule_writer;
 	private String calendar_schedule_writer_name;
 	private LocalDateTime calendar_schedule_start_date;
 	private LocalDateTime calendar_schedule_end_date;
