+++ src/main/java/com/codenal/employee/repository/EmployeeRepository.java
@@ -5,5 +5,5 @@ import com.codenal.employee.domain.Employee;
 
 public interface EmployeeRepository extends JpaRepository<Employee, Long> {
     // 필요시 커스텀 쿼리 추가 가능
-    Employee findByEmpId(int empId);
+    Employee findByEmpId(Long empId);
 }
